kafka-producer/requirements.txt
kafka-python
python-dotenv
requests
pytz
schedule
beautifulsoup4
lxml
flask
flask_cors